Partial integration and refactoring to code base
CHANGES:
weapoons.h->Weapon.h
consumable.h->Consumable.h
item.h->Item.h
event.h->Event.h

MODIFICATIONS:
- Added necessary header files to:
1. CombatEv
2. Enemy.h
3. EnemyData.h(Added: class Weapon; for private Weapon var)

NOTABLE:
- Only compiling character classes in CMAKE as of now
- Need to look at Enemy class and update equip weapon to follow suite with current plans for weapons

+++ header/Consumable.h
@@ -2,7 +2,7 @@
 #define CONSUMBALE_H
 
 #include <iostream>
-#include "item.h"
+#include "Item.h"
 
 using std::string;
 
@@ -12,7 +12,7 @@ class Consumable : public Item {
         string type;
     public:
         Consumable(string, string, float);
-        float useItem();
+        int useItem();
         void description();
 };
